refactor: change imports in tests to make node happy
