+++ packages/common/src/capabilities/_internal/index.ts
@@ -1,0 +1,1 @@
+export * from "./checkCapabilityAccess";
